fix: registerSchema 객체화
