+++ src/contexts/TechStacksContext.ts
@@ -1,4 +1,4 @@
-import type { TechStack } from "@/types/commonTypes";
+import type { TechStack } from "@/types/techStackTypes";
 import { createContext } from "react";
 
 export const TechStacksContext = createContext<TechStack[]>([]);
