+++ src/types/apiTypes.ts
@@ -4,9 +4,3 @@ export type ApiErrorResponse = {
   error: string;
   path: string;
 };
-
-export type TechStack = {
-  id: number;
-  name: string;
-  level?: number;
-};
