+++ src/types/techStackTypes.ts
@@ -1,0 +1,5 @@
+export type TechStack = {
+  id: number;
+  name: string;
+  level?: number;
+};
